requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.1.0
Django==5.2.1
django-allauth==65.9.0
django-cors-headers==4.7.0
django-csp==3.7
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.27.1
drf-nested-routers==0.94.2
gunicorn==21.2.0
idna==3.10
Markdown==3.8
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
